INDY-1879: fix onPoolMembershipChange and add tests
