+++ plenum/test/pool_transactions/conftest.py
@@ -1,5 +1,7 @@
 import pytest
 
+from plenum.common.constants import NODE
+from plenum.common.txn_util import get_type
 from plenum.common.util import randomString
 from plenum.test.test_node import checkNodesConnected, TestNode
 from plenum.test.pool_transactions.helper import \
@@ -43,3 +45,13 @@ def sdk_node_theta_added(looper,
     looper.run(checkNodesConnected(txnPoolNodeSet))
     sdk_pool_refresh(looper, sdk_pool_handle)
     return new_steward_wallet, new_node
+
+
+@pytest.fixture()
+def pool_node_txns(poolTxnData):
+    node_txns = []
+    for txn in poolTxnData["txns"]:
+        if get_type(txn) == NODE:
+            node_txns.append(txn)
+    return node_txns
+
